gateway/pyproject.toml
[project]
name = "gsma-open-gateway"
version = "0.1.0"
description = "IT's GSMA Open Gateway implementation"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.115.11",
    "geopy>=2.4.1",
    "pydantic-settings>=2.8.1",
    "redis>=5.2.1",
    "requests>=2.32.3",
]

[tool.mypy]
strict = true

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.15.0",
    "pytest>=8.3.5",
    "ruff>=0.11.0",
]
